+++ internal/codeintel/enqueuer/indexer_name.go
@@ -1,4 +1,4 @@
-package server
+package enqueuer
 
 import (
 	"bufio"